% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{class}
\name{fit_summaries-class}
\alias{fit_summaries-class}
\title{An S4 class to represent summaries of fit}
\description{
An S4 class to represent summaries of fit
}
\section{Slots}{

\describe{
\item{\code{pps}}{data frame with predictions for response proportions}

\item{\code{RTs}}{data frame with predictions for RTs}
}}

