\name{PMDC-package}
\alias{PMDC-package}
\alias{PMDC}
\docType{package}
\title{
\packageTitle{PMDC}
}
\description{
\packageDescription{PMDC}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{PMDC}
\packageIndices{PMDC}
~~ An overview of how to use the package, including ~~
~~ the most important functions ~~
}
\author{
\packageAuthor{PMDC}

Maintainer: \packageMaintainer{PMDC}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, ~~
~~ from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
